fixed some merge issue
