+++ features/step_definitions/map_steps.rb
@@ -19,3 +19,22 @@ end
 Then('I should see myself be on the Florida state page') do
   expect(page).to have_current_path('/state/FL')
 end
+Then /^(?:|I )should see "([^"]*)"$/ do |text|
+  if page.respond_to? :should
+    page.should have_content(text)
+  else
+    assert page.has_content?(text)
+  end
+end
+
+Then /^(?:|I )should not see "([^"]*)"$/ do |text|
+  if page.respond_to? :should
+    page.should have_no_content(text)
+  else
+    assert page.has_no_content?(text)
+  end
+end
+
+When /^(?:|I )go to (.+)$/ do |page_name|
+     visit path_to(page_name)
+end
